+++ CoCowork.DB/Tables/Place.sql
@@ -2,7 +2,7 @@
 (
 	[Id] INT NOT NULL PRIMARY KEY IDENTITY, 
     [MiniOfficeId] INT NULL, 
-    [PricePerDay] INT NOT NULL, 
-    [PriceFixedPerDay] INT NOT NULL, 
+    [PricePerDay] DECIMAL(10, 2) NOT NULL, 
+    [PriceFixedPerDay] DECIMAL(10, 2) NOT NULL, 
     CONSTRAINT [FK_Place_ToMiniOffice] FOREIGN KEY ([MiniOfficeId]) REFERENCES [dbo].[MiniOffice]([Id]),
 )
