+++ CoCowork.DB/Tables/PlaceOrder.sql
@@ -5,7 +5,7 @@
     [OrderId] INT NOT NULL, 
     [StartDate] DATETIME NOT NULL, 
     [EndDate] DATETIME NOT NULL, 
-    [SubtotalPrice] INT NOT NULL, 
+    [SubtotalPrice] DECIMAL(10, 2) NOT NULL, 
     CONSTRAINT [FK_PlaceOrder_ToPlace] FOREIGN KEY ([PlaceId]) REFERENCES [dbo].[Place]([Id]),
 
 )
